Enum gfx_glyph::Layout

pub enum Layout<L: LineBreaker> {
    SingleLine {
        line_breaker: L,
        h_align: HorizontalAlign,
        v_align: VerticalAlign,
    },
    Wrap {
        line_breaker: L,
        h_align: HorizontalAlign,
        v_align: VerticalAlign,
    },
}

Built-in GlyphPositioner implementations.

Takes generic LineBreaker to indicate the wrapping style. See BuiltInLineBreaker.

Example

let layout = Layout::default().h_align(HorizontalAlign::Right);

Variants
SingleLine

Renders a single line from left-to-right according to the inner alignment. Hard breaking will end the line, partially hitting the width bound will end the line.

Fields of SingleLine

line_breaker: L

h_align: HorizontalAlign

v_align: VerticalAlign

Wrap

Renders multiple lines from left-to-right according to the inner alignment. Hard breaking characters will cause advancement to another line. A characters hitting the width bound will also cause another line to start.

Fields of Wrap

line_breaker: L

h_align: HorizontalAlign

v_align: VerticalAlign

Methods

impl Layout<BuiltInLineBreaker>

pub fn default_single_line() -> Self

pub fn default_wrap() -> Self

impl<L: LineBreaker> Layout<L>

pub fn h_align(self, h_align: HorizontalAlign) -> Self

Returns an identical Layout but with the input h_align

pub fn v_align(self, v_align: VerticalAlign) -> Self

Returns an identical Layout but with the input v_align

pub fn line_breaker<L2: LineBreaker>(self, line_breaker: L2) -> Layout<L2>

Returns an identical Layout but with the input line_breaker

impl<L: LineBreaker> Layout<L>

pub fn calculate_glyphs_and_leftover<'a>(
&self,
font_map: &HashMap<FontId, Font>,
section: &SectionGlyphInfo<'a>
) -> (Vec<GlyphedSectionText>, Option<LayoutLeftover<'a>>)