Struct gfx_glyph::GlyphInfo

pub struct GlyphInfo<'a> {
    pub text: &'a str,
    pub scale: Scale,
    pub color: [f32; 4],
    pub font_id: FontId,
    // some fields omitted
}

A specialised view on a Section for the purposes of calculating glyph positions. Used by a GlyphPositioner.

See Layout for built-in positioner logic.

Fields
text: &'a str

Section text, use remaining_chars() instead in order to respect skip settings, ie in leftover payloads.

scale: Scale color: [f32; 4] font_id: FontId
Methods

impl<'a> GlyphInfo<'a>

pub fn remaining_chars(&self) -> Skip<Recompositions<Chars<'a>>>

Returns a unicode normalized char iterator, that respects the skipped chars that have already been already processed

Important traits for RemainingNormCharIndices<'a>
impl<'a> Iterator for RemainingNormCharIndices<'a> type Item = (usize, char);

pub fn remaining_char_indices(&self) -> RemainingNormCharIndices<'a>

Returns a unicode normalized (byte_index, char) iterator, that respects the skipped chars that have already been already processed

pub fn skip(&self, skip: usize) -> GlyphInfo<'a>

Returns a new GlyphInfo instance whose remaining_chars() method will skip additional chars (not bytes!)

pub fn substring(&self) -> &'a str

Returns a substring reference according the current skip value