Struct gfx_glyph::SectionGlyphInfo

pub struct SectionGlyphInfo<'a> {
    pub screen_position: (f32, f32),
    pub bounds: (f32, f32),
    pub text: Vec<GlyphInfo<'a>>,
    pub text_index: usize,
}

Fields
screen_position: (f32, f32)

Position on screen to render text, in pixels from top-left.

bounds: (f32, f32)

Max (width, height) bounds, in pixels from top-left.

text: Vec<GlyphInfo<'a>> text_index: usize
Methods

impl<'a> SectionGlyphInfo<'a>

pub fn with_info(&self, index: usize, info: GlyphInfo<'a>) -> Self

Returns a clone info that has marked text up to index as processed, and GlyphInfo at index as input info

pub fn remaining_info(&self) -> Skip<Enumerate<Iter<GlyphInfo<'a>>>>