Struct gfx_glyph::GlyphBrush

pub struct GlyphBrush<'font, R: Resources, F: Factory<R>> { /* fields omitted */ }

Object allowing glyph drawing, containing cache state. Manages glyph positioning cacheing, glyph draw caching & efficient GPU texture cache updating and re-sizing on demand.

Build using a GlyphBrushBuilder.

Example

extern crate gfx_glyph;
use gfx_glyph::Section;

let section = Section {
    text: "Hello gfx_glyph",
    ..Section::default()
};

glyph_brush.queue(section);
glyph_brush.queue(some_other_section);

glyph_brush.draw_queued(&mut gfx_encoder, &gfx_color, &gfx_depth).unwrap();

Caching behaviour

Calls to GlyphBrush::queue, GlyphBrush::pixel_bounds, GlyphBrush::glyphs calculate the positioned glyphs for a section. This is cached so future calls to any of the methods for the same section are much cheaper. In the case of GlyphBrush::queue the calculations will also be used for actual drawing.

The cache for a section will be cleared after a GlyphBrush::draw_queued call when that section has not been used since the previous draw call.

Methods

impl<'font, R: Resources, F: Factory<R>> GlyphBrush<'font, R, F>

pub fn queue_custom_layout<'a, S, G>(&mut self, section: S, custom_layout: &G) where
G: GlyphPositioner,
S: Into<Cow<'a, VariedSection<'a>>>,

Queues a section/layout to be drawn by the next call of draw_queued. Can be called multiple times to queue multiple sections for drawing.

Used to provide custom GlyphPositioner logic, if using built-in Layout simply use queue

Benefits from caching, see caching behaviour.

pub fn queue<'a, S>(&mut self, section: S) where
S: Into<Cow<'a, VariedSection<'a>>>,

Queues a section/layout to be drawn by the next call of draw_queued. Can be called multiple times to queue multiple sections for drawing.

Benefits from caching, see caching behaviour.

pub fn draw_queued<C, T, D>(
&mut self,
encoder: &mut Encoder<R, C>,
target: &RenderTargetView<R, T>,
depth_target: &DepthStencilView<R, D>
) -> Result<(), String> where
C: CommandBuffer<R>,
T: RenderFormat,
D: DepthFormat,

Draws all queued sections onto a render target, applying a position transform (e.g. a projection). See queue.

Trims the cache, see caching behaviour.

pub fn draw_queued_with_transform<C, T, D>(
&mut self,
transform: [[f32; 4]; 4],
encoder: &mut Encoder<R, C>,
target: &RenderTargetView<R, T>,
depth_target: &DepthStencilView<R, D>
) -> Result<(), String> where
C: CommandBuffer<R>,
T: RenderFormat,
D: DepthFormat,

Draws all queued sections onto a render target, applying a position transform (e.g. a projection). See queue.

Trims the cache, see caching behaviour.

pub fn fonts(&self) -> &HashMap<FontId, Font<'font>>

Returns FontId -> Font map of available fonts.

pub fn add_font_bytes<'a: 'font, B: Into<SharedBytes<'a>>>(
&mut self,
font_data: B
) -> FontId

Adds an additional font to the one(s) initially added on build.

Returns a new FontId to reference this font.

Example

extern crate gfx_glyph;
use gfx_glyph::{Section, GlyphBrushBuilder};

// dejavu is built as default `FontId(0)`
let dejavu: &[u8] = include_bytes!("../examples/DejaVuSans.ttf");
let mut glyph_brush = GlyphBrushBuilder::using_font_bytes(dejavu)
    .build(gfx_factory.clone());

// some time later, add another font referenced by a new `FontId`
let open_sans_italic: &[u8] = include_bytes!("../examples/OpenSans-Italic.ttf");
let open_sans_italic_id = glyph_brush.add_font_bytes(open_sans_italic);

pub fn add_font<'a: 'font>(&mut self, font_data: Font<'a>) -> FontId

Adds an additional font to the one(s) initially added on build.

Returns a new FontId to reference this font.