Trait gfx_glyph::GlyphPositioner

pub trait GlyphPositioner: Hash {
    fn calculate_glyphs<'font>(
        &self, 
        _: &FontMap<'font>, 
        section: &VariedSection
    ) -> Vec<(PositionedGlyph<'font>, [f32; 4], FontId)>;

    fn bounds_rect(&self, section: &VariedSection) -> Rect<f32>;
}

Logic to calculate glyph positioning based on Font and VariedSection

Required Methods

Important traits for Vec<u8>
impl Write for Vec<u8>

fn calculate_glyphs<'font>(
&self,
_: &FontMap<'font>,
section: &VariedSection
) -> Vec<(PositionedGlyph<'font>, [f32; 4], FontId)>

Calculate a sequence of positioned glyphs to render. Custom implementations should always return the same result when called with the same arguments. If not consider disabling cache_glyph_positioning.

fn bounds_rect(&self, section: &VariedSection) -> Rect<f32>

Return a screen rectangle according to the requested render position and bounds appropriate for the glyph layout.